backend/requirements.txt
fastapi[standard]
boto3
python-dotenv
pytesseract
google-genai
torch
torchvision
requests